Move openhands.server.user_auth to openhands.app_server.user_auth
- Move user_auth directory from openhands/server to openhands/app_server
- Update all imports across the codebase (openhands, enterprise, tests)
- Update default user_auth_class path in server_config.py
- Update documentation examples in import_utils.py
