+++ enterprise/integrations/jira_dc/jira_dc_manager.py
@@ -38,7 +38,7 @@ from openhands.server.types import (
     MissingSettingsError,
     SessionExpiredError,
 )
-from openhands.server.user_auth.user_auth import UserAuth
+from openhands.app_server.user_auth.user_auth import UserAuth
 from openhands.utils.http_session import httpx_verify_option
