+++ enterprise/integrations/slack/slack_view.py
@@ -36,7 +36,7 @@ from openhands.app_server.services.injector import InjectorState
 from openhands.app_server.user.specifiy_user_context import USER_CONTEXT_ATTR
 from openhands.core.logger import openhands_logger as logger
 from openhands.sdk import TextContent
-from openhands.server.user_auth.user_auth import UserAuth
+from openhands.app_server.user_auth.user_auth import UserAuth
 from openhands.utils.async_utils import GENERAL_TIMEOUT
 
 # =================================================
