+++ enterprise/integrations/types.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from integrations.models import Message
 
     from openhands.app_server.integrations.provider import PROVIDER_TOKEN_TYPE
-    from openhands.server.user_auth.user_auth import UserAuth
+    from openhands.app_server.user_auth.user_auth import UserAuth
 
 
 class GitLabResourceType(Enum):
