+++ enterprise/server/auth/saas_user_auth.py
@@ -43,7 +43,7 @@ from openhands.app_server.integrations.provider import (
 from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.app_server.settings.settings_models import Settings
 from openhands.app_server.settings.settings_store import SettingsStore
-from openhands.server.user_auth.user_auth import AuthType, UserAuth
+from openhands.app_server.user_auth.user_auth import AuthType, UserAuth
 
 token_manager = TokenManager()
