+++ enterprise/server/email_validation.py
@@ -5,7 +5,7 @@ Email domain validation utilities for enterprise endpoints.
 from fastapi import Depends, HTTPException, Request, status
 
 from openhands.core.logger import openhands_logger as logger
-from openhands.server.user_auth import get_user_auth, get_user_id
+from openhands.app_server.user_auth import get_user_auth, get_user_id
 
 
 async def get_admin_user_id(
