+++ enterprise/server/middleware.py
@@ -17,7 +17,7 @@ from server.utils.url_utils import get_cookie_domain, get_cookie_samesite
 
 from openhands.core.logger import openhands_logger as logger
 from openhands.server.shared import config
-from openhands.server.user_auth.user_auth import AuthType, UserAuth, get_user_auth
+from openhands.app_server.user_auth.user_auth import AuthType, UserAuth, get_user_auth
 
 
 class SetAuthCookieMiddleware:
