+++ enterprise/server/routes/billing.py
@@ -21,7 +21,7 @@ from storage.subscription_access import SubscriptionAccess
 from storage.user_store import UserStore
 
 from openhands.app_server.config import get_global_config
-from openhands.server.user_auth import get_user_id
+from openhands.app_server.user_auth import get_user_id
 
 stripe.api_key = STRIPE_API_KEY
 billing_router = APIRouter(prefix='/api/billing', tags=['Billing'])
