+++ enterprise/server/routes/oauth_device.py
@@ -11,7 +11,7 @@ from storage.api_key_store import ApiKeyStore
 from storage.device_code_store import DeviceCodeStore
 
 from openhands.core.logger import openhands_logger as logger
-from openhands.server.user_auth import get_user_id
+from openhands.app_server.user_auth import get_user_id
 
 # ---------------------------------------------------------------------------
 # Constants
