+++ enterprise/tests/unit/integrations/jira_dc/conftest.py
@@ -17,7 +17,7 @@ from storage.jira_dc_user import JiraDcUser
 from storage.jira_dc_workspace import JiraDcWorkspace
 
 from openhands.app_server.integrations.service_types import ProviderType, Repository
-from openhands.server.user_auth.user_auth import UserAuth
+from openhands.app_server.user_auth.user_auth import UserAuth
 
 
 @pytest.fixture
