+++ enterprise/tests/unit/server/routes/test_orgs.py
@@ -45,7 +45,7 @@ from server.routes.orgs import (
 from storage.org import Org
 
 from openhands.sdk.settings import AgentSettings, ConversationSettings
-from openhands.server.user_auth import get_user_id
+from openhands.app_server.user_auth import get_user_id
 
 # Test user ID constant (must be a valid UUID string)
 TEST_USER_ID = str(uuid.uuid4())
