+++ openhands/app_server/settings/settings_router.py
@@ -29,18 +29,18 @@ from openhands.app_server.settings.settings_models import (
     Settings,
 )
 from openhands.app_server.settings.settings_store import SettingsStore
-from openhands.app_server.utils.dependencies import get_dependencies
-from openhands.core.logger import openhands_logger as logger
-from openhands.sdk.llm import LLM
-from openhands.sdk.settings import ConversationSettings
-from openhands.server.shared import config
-from openhands.server.user_auth import (
+from openhands.app_server.user_auth import (
     get_provider_tokens,
     get_secrets_store,
     get_user_id,
     get_user_settings,
     get_user_settings_store,
 )
+from openhands.app_server.utils.dependencies import get_dependencies
+from openhands.core.logger import openhands_logger as logger
+from openhands.sdk.llm import LLM
+from openhands.sdk.settings import ConversationSettings
+from openhands.server.shared import config
 from openhands.utils.llm import (
     get_provider_api_base,
     is_openhands_model,
