+++ openhands/app_server/user_auth/default_user_auth.py
@@ -16,8 +16,8 @@ from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.app_server.secrets.secrets_store import SecretsStore
 from openhands.app_server.settings.settings_models import Settings
 from openhands.app_server.settings.settings_store import SettingsStore
+from openhands.app_server.user_auth.user_auth import UserAuth
 from openhands.server import shared
-from openhands.server.user_auth.user_auth import UserAuth
 
 
 @dataclass
