+++ tests/unit/server/test_openapi_schema_generation.py
@@ -13,8 +13,8 @@ from openhands.app_server.secrets.secrets_models import Secrets
 from openhands.app_server.secrets.secrets_store import SecretsStore
 from openhands.app_server.settings.file_settings_store import FileSettingsStore
 from openhands.app_server.settings.settings_store import SettingsStore
+from openhands.app_server.user_auth.user_auth import UserAuth
 from openhands.server.app import app
-from openhands.server.user_auth.user_auth import UserAuth
 
 
 class MockUserAuth(UserAuth):
@@ -67,7 +67,7 @@ def test_client():
     # Create a test client
     with (
         patch(
-            'openhands.server.user_auth.user_auth.UserAuth.get_instance',
+            'openhands.app_server.user_auth.user_auth.UserAuth.get_instance',
             return_value=MockUserAuth(),
         ),
         patch(
